Pulling up some methods
