+++ app/models/user_mode.rb
@@ -1,5 +1,21 @@
-module UserMode
+class UserMode
   extend ConfigurableGlobal
+
+  def protect_from_forgery(controller)
+    controller.protect_from_forgery with: :exception
+  end
+
+  def logout_redirection_url(controller)
+    controller.after_logout_redirection_url
+  end
+
+  def html_badge
+    ''
+  end
+
+  def auth_init_partial
+    'layouts/auth_partials/null_partial.html.erb'
+  end
 
   def self.get_current
     auth_strategy = Rails.configuration.auth_provider.downcase
