+++ app/models/user_mode/auth0_auth_strategy.rb
@@ -1,4 +1,4 @@
-class UserMode::Auth0AuthStrategy < UserMode::MultiUser
+class UserMode::Auth0AuthStrategy < UserMode
   def set_auth_provider(omniauth)
     omniauth.provider :auth0,
                       Rails.configuration.auth0_client_id,
@@ -11,14 +11,6 @@ class UserMode::Auth0AuthStrategy < UserMode::MultiUser
     'href="#" onclick="window.signin();"'
   end
 
-  def protect_from_forgery(controller)
-    controller.protect_from_forgery with: :exception
-  end
-
-  def logout_redirection_url(controller)
-    controller.after_logout_redirection_url
-  end
-
   def auth_init_partial
     'layouts/auth_partials/auth0.init.html.erb'
   end
@@ -26,6 +18,4 @@ class UserMode::Auth0AuthStrategy < UserMode::MultiUser
   def html_badge
     '<a href="https://auth0.com/" target="_blank"><img height="40" alt="JWT Auth for open source projects" src="//cdn.auth0.com/oss/badges/a0-badge-light.png"/></a>'.html_safe
   end
-
-
 end
