--- app/models/user_mode/multi_user.rb
@@ -1,2 +1,0 @@
-class UserMode::MultiUser
-end
