+++ app/models/user_mode/single_user.rb
@@ -1,4 +1,5 @@
-class UserMode::SingleUser
+class UserMode::SingleUser < UserMode
+
   def set_auth_provider(omniauth)
     omniauth.provider :developer
   end
@@ -7,19 +8,6 @@ class UserMode::SingleUser
     "href='/auth/developer'"
   end
 
-  def protect_from_forgery(controller)
-    # not needed!
-  end
-
-  def logout_redirection_url(controller)
-    controller.after_logout_redirection_url
-  end
-
-  def html_badge
-    ''
-  end
-
-  def auth_init_partial
-    'layouts/auth_partials/null_partial.html.erb'
+  def protect_from_forgery(_controller)
   end
 end
